requirements.txt
gradio>=4.0.0
easyocr>=1.7.0
opencv-python-headless>=4.8.0
python-docx>=1.0.0
Pillow>=10.0.0
numpy>=1.24.0
torch>=2.6.0
transformers>=4.30.0
sentencepiece>=0.1.99
scipy>=1.10.0
accelerate>=0.20.0
